refactor: Request, Response 패키지 수정
