+++ src/main/java/racingcar/consolegame/ui/OutputView.java
@@ -1,6 +1,6 @@
 package racingcar.consolegame.ui;
 
-import racingcar.dto.RacingResponse;
+import racingcar.controller.RacingResponse;
 import racingcar.model.Car;
 
 /**
