+++ src/main/java/racingcar/controller/RacingGameRequest.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.controller;
 
 import java.beans.ConstructorProperties;
