+++ src/main/java/racingcar/controller/RacingcarController.java
@@ -8,8 +8,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.dto.RacingGameRequest;
-import racingcar.dto.RacingResponse;
 import racingcar.service.RacingcarService;
 
 @RestController
